基本完成了后台操作

+++ src/main/java/com/sherlockvarious/blog/Controller/BlogController.java
@@ -6,6 +6,7 @@ import com.sherlockvarious.blog.service.BlogService;
 import com.sherlockvarious.blog.service.Blog_TagsService;
 import com.sherlockvarious.blog.service.TagService;
 import com.sherlockvarious.blog.service.TypeService;
+import com.sun.mail.imap.protocol.ID;
 import com.sun.org.apache.xpath.internal.operations.Mod;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,6 +18,8 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -51,12 +54,34 @@ public class BlogController {
     }
 
     @RequestMapping("/admin/blog/publish")
-    public String publishBlog(Blog blog, String tagIds, RedirectAttributes attributes, HttpSession session) {
-        User user = (User) session.getAttribute("user");
-        blog.setUserId(user.getId());
+    public String publishBlog(Blog blog, String tagIds, RedirectAttributes attributes, HttpSession session,Model model) {
 
-        if (blogService.saveNewBlog(blog) == true) {
-            blog_tagsService.addNewRelationship(blog.getId(),tagIds);
+        if(blog.getTypeId()==null||"".equals(tagIds)){
+            model.addAttribute("message", "类型不能为空");
+            model.addAttribute("types", typeService.listAllType());
+            model.addAttribute("tags", tagService.listAllTags());
+            return "admin/blogs-publish";
+        }
+        if (blog.getId() == null) {
+            User user = (User) session.getAttribute("user");
+            blog.setUserId(user.getId());
+
+            if (blogService.saveNewBlog(blog) == true) {
+                blog_tagsService.addNewRelationship(blog.getId(), tagIds);
+                attributes.addFlashAttribute("message", "操作成功");
+                return "redirect:/page/admin/blogs";
+            }
+
+            attributes.addFlashAttribute("message", "操作失败");
+
+        }else {
+            blog.setUpdateTime(new Date());
+            blogService.updateBlog(blog);
+            System.out.println(blogService.findTags(blog.getId()));
+            if(!(blogService.findTags(blog.getId()).equals(tagIds))){
+                blog_tagsService.deleteByBlogId(blog.getId());
+                blog_tagsService.addNewRelationship(blog.getId(), tagIds);
+            }
         }
 
         return "redirect:/page/admin/blogs";
@@ -78,6 +103,10 @@ public class BlogController {
 
     @RequestMapping("/admin/blogs/delete/{id}")
     public String delete(@PathVariable int id, RedirectAttributes attributes) {
+        if (!blog_tagsService.deleteByBlogId(id)) {
+            attributes.addFlashAttribute("message", "删除失败");
+            return "redirect:/page/admin/blogs";
+        }
         if (blogService.deleteById(id)) {
             attributes.addFlashAttribute("message", "删除成功");
         } else {
@@ -85,6 +114,20 @@ public class BlogController {
         }
 
         return "redirect:/page/admin/blogs";
+    }
+
+    @RequestMapping("/admin/blogs/edit/{id}")
+    public String edit(@PathVariable int id, Model model) {
+        model.addAttribute("types", typeService.listAllType());
+        model.addAttribute("tags", tagService.listAllTags());
+
+        Blog blog = blogService.getBlog(id);
+
+        blog.setTagIds(blogService.findTags(id));
+
+        model.addAttribute("blog", blog);
+
+        return "admin/blogs-publish";
     }
 
     @GetMapping("/page/admin/blogs-publish")
